Faster modals + movile dept search + stensal banner
For mobile: a larger dept search scroll.
https://github.com/the-rango/AntAlmanac/issues/61

It turns out that the almanac graph modals were a bottleneck, so I restructured the code for faster rendering. https://github.com/the-rango/AntAlmanac/issues/26
