+++ src/components/SearchForm/SearchForm.js
@@ -1,4 +1,5 @@
 import DeptSearchBar from './DeptSearchBar/DeptSearchBar';
+import MobileDeptSelector from './DeptSearchBar/MobileDeptSelector';
 import GESelector from './GESelector/GESelector';
 import TermSelector from './TermSelector';
 import CourseCodeSearchBar from './CourseCodeSearchBar';
@@ -136,6 +137,10 @@ class SearchForm extends Component {
     else this.setState({ dept: dept.value, label: dept.label });
   };
 
+  setDeptMobile = (dept) => {
+    this.setState({ dept: dept });
+  };
+
   handleAdvancedSearchChange = (advancedSearchState) => {
     this.setState(advancedSearchState);
   };
@@ -182,7 +187,14 @@ class SearchForm extends Component {
         </div>
 
         <div className={classes.margin}>
-          <DeptSearchBar dept={this.state.label} setDept={this.setDept} />
+          {isMobile ? (
+            <MobileDeptSelector
+              dept={this.state.dept}
+              setDept={this.setDeptMobile}
+            />
+          ) : (
+            <DeptSearchBar dept={this.state.label} setDept={this.setDept} />
+          )}
           <CourseNumberSearchBar
             //Places CourseNumberSearchBar object next to DeptSearchBar object
             onAdvancedSearchChange={this.handleAdvancedSearchChange}
